Checkpoint packaging and related changes.
