+++ Gemfile
@@ -7,6 +7,6 @@ gem "byebug", "~> 11.1.3"
 gem "dotenv", "~> 2.8.1"
 gem "rake", "~> 13.0"
 gem "rspec", "~> 3.12"
-gem "rubocop", "~> 1.41.1"
+gem "rubocop", "~> 1.42.0"
 gem "vcr", "~> 6.1.0"
 gem "webmock", "~> 3.18.1"
