+++ lib/ruby/openai/client_api/models.rb
@@ -1,0 +1,17 @@
+module OpenAI
+  module ClientApi
+    class Models
+      def initialize(client: nil)
+        @client = client
+      end
+
+      def list
+        @client.get(path: "/models")
+      end
+
+      def retrieve(id:)
+        @client.get(path: "/models/#{id}")
+      end
+    end
+  end
+end
