+++ lib/ruby/openai/version.rb
@@ -1,5 +1,3 @@
-module Ruby
-  module OpenAI
-    VERSION = "3.0.2".freeze
-  end
+module OpenAI
+  VERSION = "3.0.2".freeze
 end
