+++ spec/openai/client/client_spec.rb
