+++ spec/openai/client/completions/codex_spec.rb
