+++ spec/openai/client/completions/gpt3_spec.rb
