+++ spec/openai/client/edits_spec.rb
