+++ spec/openai/client/embeddings_spec.rb
