+++ spec/openai/client/files_spec.rb
