+++ spec/openai/client/images_spec.rb
