+++ spec/openai/client/models_spec.rb
