+++ spec/openai/client/moderations_spec.rb
