+++ spec/openai_spec.rb
@@ -1,6 +1,6 @@
-RSpec.describe Ruby::OpenAI do
+RSpec.describe OpenAI do
   it "has a version number" do
-    expect(Ruby::OpenAI::VERSION).not_to be nil
+    expect(OpenAI::VERSION).not_to be nil
   end
 
   describe "#configure" do
@@ -9,7 +9,7 @@ RSpec.describe Ruby::OpenAI do
     let(:organization_id) { "def456" }
 
     before do
-      Ruby::OpenAI.configure do |config|
+      OpenAI.configure do |config|
         config.access_token = access_token
         config.api_version = api_version
         config.organization_id = organization_id
@@ -17,16 +17,16 @@ RSpec.describe Ruby::OpenAI do
     end
 
     it "returns the config" do
-      expect(Ruby::OpenAI.configuration.access_token).to eq(access_token)
-      expect(Ruby::OpenAI.configuration.api_version).to eq(api_version)
-      expect(Ruby::OpenAI.configuration.organization_id).to eq(organization_id)
+      expect(OpenAI.configuration.access_token).to eq(access_token)
+      expect(OpenAI.configuration.api_version).to eq(api_version)
+      expect(OpenAI.configuration.organization_id).to eq(organization_id)
     end
 
     context "without an access token" do
       let(:access_token) { nil }
 
       it "raises an error" do
-        expect { OpenAI::Client.new.completions }.to raise_error(Ruby::OpenAI::ConfigurationError)
+        expect { OpenAI::Client.new.completions }.to raise_error(OpenAI::ConfigurationError)
       end
     end
   end
